+++ tests/test_ht1621.py
@@ -9,6 +9,7 @@ Tests for the :py:class:`luma.lcd.device.ht1621` device.
 
 from luma.lcd.device import ht1621
 from luma.core.virtual import sevensegment
+from helpers import Mock  # noqa: F401
 
 
 class MockHT1621:
@@ -136,12 +137,14 @@ def setup_function(function):
 
 
 def test_init_6x8():
-    ht1621(gpio)
+    ht1621(gpio, serial_interface=Mock())
 
     assert gpio.get_pin_states() == {
         11: gpio.OUT,
         10: gpio.OUT,
-        8: gpio.OUT}
+        8: gpio.OUT,
+        18: gpio.OUT
+    }
 
     assert gpio.get_data() == [
         gpio.command(0x30),  # Internal RC oscillator @ 256KHz
@@ -153,7 +156,8 @@ def test_init_6x8():
         gpio.data(0),        #
         gpio.data(0),        #
         gpio.data(0),        #
-        gpio.command(0x06)]  # Display On
+        gpio.command(0x06)   # Display On
+    ]
 
 
 def test_cleanup():